refactor(api): move /api/submissions/* to app dir
