+++ src/lib/api/schema/searchParams.ts
@@ -1,0 +1,7 @@
+export function paramsToObject(search: URLSearchParams) {
+  const params: Record<string, string> = {}
+  for (const [key, value] of search.entries()) {
+    params[key] = value
+  }
+  return params
+}
